Radiance MacApp projects. Ported Radiance to gcc LP64. Fixed horrible issue in StringBase where the vprintf family of functions apparently modify the va_list argument on the LP64 compiler, using va_copy on GCC to fix this. This code is untested on windows and will probably break.

+++ Source/Engine/Lua/LuaRuntime.cpp
@@ -147,7 +147,7 @@ struct LuaPools
 				AddrSize chunkSize = BasePoolSize<<i;
 
 				char sz[64];
-				sprintf(sz, "lua_pool_%d", chunkSize);
+				sprintf(sz, "lua_pool_%d", (int)chunkSize);
 
 				s_pools[i].Create(
 					ZLuaRuntime,
